requirements.txt
streamlit
faiss-cpu
sentence-transformers
pandas
numpy
numexpr==2.8.4
bottleneck==1.3.7
